lets add a tenant field

+++ lib/phil_columns/seed/schema_seed.ex
@@ -8,6 +8,7 @@ defmodule PhilColumns.Seed.SchemaSeed do
   @primary_key false
   schema "schema_seeds" do
     field :version, :integer
+    field :tenant, :string
     timestamps updated_at: false
   end
 
@@ -18,16 +19,16 @@ defmodule PhilColumns.Seed.SchemaSeed do
     create_seeds_table(adapter, repo)
   end
 
-  def seeded_versions(repo) do
-    repo.all from(p in {get_source(repo), __MODULE__}, select: p.version), @opts
+  def seeded_versions(repo, tenant) do
+    repo.all from(p in {get_source(repo), __MODULE__}, select: p.version, where: p.tenant == ^tenant), @opts
   end
 
-  def up(repo, version) do
-    repo.insert! %__MODULE__{version: version}, @opts
+  def up(repo, version, tenant) do
+    repo.insert! %__MODULE__{version: version, tenant: tenant}, @opts
   end
 
-  def down(repo, version) do
-    repo.delete_all from(p in __MODULE__, where: p.version == ^version), @opts
+  def down(repo, version, tenant) do
+    repo.delete_all from(p in __MODULE__, where: p.version == ^version and p.tenant == ^tenant), @opts
   end
 
   def get_source(repo) do
@@ -41,7 +42,8 @@ defmodule PhilColumns.Seed.SchemaSeed do
     # DDL queries do not log, so we do not need to pass log: false here.
     adapter.execute_ddl(repo,
       {:create_if_not_exists, table, [
-        {:add, :version, :bigint, primary_key: true},
+          {:add, :version, :bigint, primary_key: true},
+          {:add, :tenant, :string, default: "main"},
         {:add, :inserted_at, :naive_datetime, []}]}, @opts)
   end
 end
